fastapi-todo/requirements.txt
fastapi==0.95.2
uvicorn==0.22.0
python-jose==3.3.0
passlib==1.7.4
python-multipart==0.0.6
psycopg2-binary==2.9.10
sqlalchemy==2.0.15
pytest==7.4.0
alembic==1.11.1
pydantic==1.10.7
python-dotenv==1.0.0